python_app/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
anthropic
supabase
httpx
python-dotenv
